+++ src/components/MainDescription/MainDescription.tsx
